java to kotlin

+++ app/src/main/java/com/codesroots/osamaomar/shopgate/presentationn/screens/feature/home/cartfragment/adapter/CartAdapter.java
@@ -60,15 +60,15 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.ViewHolder>  {
             {
                 if (PreferenceHelper.getCurrencyValue()>0)
                     holder.price.setText(getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(),
-                            dataBeans.get(position).getCurrent_price() )*PreferenceHelper.getCurrencyValue()+" "+PreferenceHelper.getCurrency());
+                            (float) dataBeans.get(position).getCurrent_price())*PreferenceHelper.getCurrencyValue()+" "+PreferenceHelper.getCurrency());
                 else
                     holder.price.setText(getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(),
-                            dataBeans.get(position).getCurrent_price() )  + " " + context.getText(R.string.realcoin));
+                            (float) dataBeans.get(position).getCurrent_price())  + " " + context.getText(R.string.realcoin));
 
                 //  holder.price.setText(dataBeans.get(position).getStart_price() + " " + context.getText(R.string.realcoin));
                 products.get(position).setTotal(String.valueOf(Integer.valueOf(holder.products_count.getText().toString())*
                         getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(),
-                                dataBeans.get(position).getCurrent_price())));
+                                (float) dataBeans.get(position).getCurrent_price())));
 
 
                 products.get(position).setNotice("خصم بسبب العرض رقم"+dataBeans.get(position).getProduct().getOffers().get(0).getId());
@@ -118,7 +118,7 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.ViewHolder>  {
                 if (dataBeans.get(position).getProduct().getOffers().size()>0)
                 {
                     products.get(position).setTotal(String.valueOf(Integer.valueOf(holder.products_count.getText().toString())*
-                            getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(),dataBeans.get(position).getCurrent_price())));
+                            getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(), (float) dataBeans.get(position).getCurrent_price())));
                     products.get(position).setNotice("خصم بسبب العرض رقم "+dataBeans.get(position).getProduct().getOffers().get(0).getId());
                 }
                 else
@@ -139,7 +139,7 @@ public class CartAdapter extends RecyclerView.Adapter<CartAdapter.ViewHolder>  {
                 if (dataBeans.get(position).getProduct().getOffers().size()>0)
                 {
                     products.get(position).setTotal(String.valueOf(Integer.valueOf(holder.products_count.getText().toString())*
-                            getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(),dataBeans.get(position).getCurrent_price())));
+                            getPriceAfterDiscount(dataBeans.get(position).getProduct().getOffers().get(0).getPercentage(), (float) dataBeans.get(position).getCurrent_price())));
                     products.get(position).setNotice("خصم بسبب العرض رقم "+dataBeans.get(position).getProduct().getOffers().get(0).getId());
                 }
                 else
